Day 5, Started on the account creation controller

+++ src/main/java/f25/cs157a/evergreenbank/SavingsAccount.java
@@ -8,7 +8,7 @@ public class SavingsAccount {
     private double balance;
     private double interest;
 
-    public SavingsAccount(int routingNumber, int balance, double interest) {
+    public SavingsAccount(int routingNumber, double balance, double interest) {
         this.routingNumber = routingNumber;
         this.balance = 1000;
         this.interest = 0.01;
